Create the Code
Created the code: Import urllib, BeautifulSoup, and ssl. Prompt the user for the URL and if the user does not add a url (<1 character in length), use the default file in the code. Open and read the URL. Parse the data and give as a tree. Find all ‘comment’ tags below ‘comments’. Iterate through the data, finding the ‘count’ tag, count each ‘count’, and add the number (as an integer) to the sum. Print the final count and sum.